Some changes

+++ db/migrate/20151002014112_create_professors.rb
@@ -2,7 +2,7 @@ class CreateProfessors < ActiveRecord::Migration
   def change
     create_table :professors do |t|
       t.string :university, :last_name, :name, :title
-      t.decimal :salary, :taxable_benefits
+      t.float :salary, :taxable_benefits
       t.timestamps
     end
   end
